+++ packages/fictoan-docs/src/components/Footer/Footer.tsx
@@ -5,10 +5,10 @@ import Link from "next/link";
 import React from "react";
 
 // UI ==================================================================================================================
-import { Footer, Row, Portion, Text, Heading6, Divider } from "fictoan-react";
+import { Footer, Row, Portion, Text, Heading4, Heading6, Divider, Div } from "fictoan-react";
 
 // ASSETS ==============================================================================================================
-import FictoanLogo from "../../assets/images/fictoan-logo.svg";
+import FictoanLogo from "../../assets/images/fictoan-logo-new.svg";
 import GithubIcon from "../../assets/icons/github.svg";
 
 // STYLES ==============================================================================================================
@@ -26,12 +26,13 @@ export const SiteFooter = () => {
             <Row horizontalPadding="medium" marginBottom="none">
                 {/* BRANDING */}
                 <Portion desktopSpan="half">
-                    <div className="footer-logo">
+                    <Div className="footer-logo">
                         <FictoanLogo />
-                    </div>
-                    <Text marginTop="nano" size="small">
+                    </Div>
+
+                    <Heading4 marginTop="nano">
                         A design-centric UI framework with plain-English props
-                    </Text>
+                    </Heading4>
                 </Portion>
 
                 {/* DOCUMENTATION LINKS //////////////////////////////////////////////////////////////////////////// */}